+++ CHANGELOG.md
@@ -27,6 +27,14 @@
   [#1590](https://github.com/reupen/columns_ui/pull/1590),
   [#1591](https://github.com/reupen/columns_ui/pull/1591)]
 
+- The built-in spectrum analyser visualisation now renders in a background
+  thread, allowing rendering to continue when the main thread is blocked (such
+  as when interacting with parts of the main window title bar).
+  [[#1608](https://github.com/reupen/columns_ui/pull/1608)]
+
+  Improvements were also made to allow the visualisation to render at a higher
+  frame rate.
+
 - A main menu item, **View** › **Lock window size**, was added.
   [[#1595](https://github.com/reupen/columns_ui/pull/1595)]